% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exercise-name.R
\docType{data}
\name{animal_words}
\alias{animal_words}
\title{Generate a random name for an exercise}
\format{An object of class \code{character} of length 72.}
\usage{
animal_words
}
\description{
Generate a random name for an exercise
}
\keyword{datasets}
