% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/code_comments.R
\name{announce}
\alias{announce}
\title{Adds comments to the output of code in a learnr block.}
\usage{
announce(code, comment)
}
\arguments{
\item{code}{the code that you want to display}

\item{comment}{A character string to print as a comment}
}
\description{
Often the scaffolding put in a \code{{learnr}} exercise chunk consists of
more than one command that produces output. By default, \code{{learnr}} will
print the outputs one after the other. This can make it hard for the reader
to determine which line of code produced which output. \code{announce()} can insert comments
into the output.
}
\details{
Use \code{announce()} in either of two ways:
\enumerate{
\item As a simple function call, e.g. \code{announce("my message")} which will display
"my message" as a comment at the corresponding point in the exercise chunk output.
\item As the receiver of a piped-in command, e.g. \code{sqrt(2 + 3) \%>\% announce("my message")}. This
will print out the code on  the LHS of the pipe, along with the message as a comment.
}
}
