% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_acroscore.R
\name{do_clean_acroscore}
\alias{do_clean_acroscore}
\title{Translate an Acroscore-style question to an etude question}
\usage{
do_clean_acroscore(text)
}
\description{
This is needed only transiently, until I've worked through all
the old-style questions
}
