% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/etude_dir.R
\name{etude_dir}
\alias{etude_dir}
\title{Make a listing of an etude directory}
\usage{
etude_dir(where)
}
\arguments{
\item{where}{the name of the folder containing the etude RMD files}
}
\description{
Etude files are stored in a directory, typically "Exercises". This  function
scans through  the RMD files in such  a directory and constructs a data frame
based  on  the YAML. The data frame can be useful, for instance, for including
in a parent file using \code{etude_list()}.
}
