% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/question_compact.R
\name{etude_question}
\alias{etude_question}
\title{Custom version of \code{learnr::question()} for compact exercises.}
\usage{
etude_question(
  text,
  ...,
  type = c("auto", "single", "multiple", "learnr_radio", "learnr_checkbox",
    "learnr_text", "learnr_radio_inline", "learnr_checkbox_inline"),
  correct = "Correct!",
  incorrect = "Incorrect",
  try_again = incorrect,
  message = NULL,
  post_message = NULL,
  loading = c("**Loading:** ", text, "<br/><br/><br/>"),
  submit_button = "Submit Answer",
  try_again_button = "Try Again",
  allow_retry = FALSE,
  random_answer_order = FALSE,
  options = list()
)
}
\description{
Custom version of \code{learnr::question()} for compact exercises.
}
