% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exercise_hook.R
\name{exercise_hook}
\alias{exercise_hook}
\title{Exercise interface to knitr}
\usage{
exercise_hook(before, options, envir)
}
\arguments{
\item{before}{As with all knitr hooks, indicates whether we are before or after the chunk
is evaluated.}

\item{options}{As with all knitr hooks, the options given to the chunk}

\item{envir}{As with all knitr hooks, the environment in which the chunk will
be evaluated.}
}
\description{
A hook to include exercises in a document
}
\details{
In the startup to the document, set
`knitr::knit_hooks$set(exercise = exercise_hook)`

You don't call the hook directly. Instead, it is called by knitr
whenever the chunk option `exercise = <file_hash>` is given.
You should communicate using the chunk options, not the ar
The chunk relevant options are:
- `exercise=`: hash for exercise file
- `name=`: the identifier for the exercise to print
- `answers=`: TRUE or FALSE: show the answers?
- `package=`: the package the exercise comes from.
}
\examples{
\dontrun{
# Chunk options should be like this
exercise = "dog-flies-home", name="Exercise 36", answers = FALSE,
}
}
