% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_plots.R
\name{is_ggplot}
\alias{is_ggplot}
\title{Functions for use with gradethis exercise checking}
\usage{
is_ggplot(x)
}
\arguments{
\item{x}{input from \code{gradethis} checker, typically  \code{.result}}
}
\description{
Functions for use with gradethis exercise checking
}
