% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/new_etude_template.R
\name{new_etude_template}
\alias{new_etude_template}
\title{Make a new etude exercise}
\usage{
new_etude_template(save = TRUE, exercise_id = NULL, learnr = FALSE)
}
\arguments{
\item{save}{Logical indicating whether to save the template to a new file
and return the name of that file}

\item{exercise_id}{Root name for the exercise file (.Rmd will be added).
By default, a random, unique root name will be created using the \code{etude} naming
convention, e.g., bear-walk-dish}

\item{learnr}{Logical flag (default: FALSE) whether to use
a learnr compatible document for the exercise.}
}
\description{
Creates the boilerplate for an etude exercise.
Typically, this will be called by the RStudio addin, but
you can also run it directly when \code{save=TRUE}
}
