% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/custom_question_types.R
\name{question_ui_initialize.learnr_checkbox_inline}
\alias{question_ui_initialize.learnr_checkbox_inline}
\title{Inline checkbox}
\usage{
question_ui_initialize.learnr_checkbox_inline(question, value, ...)
}
\description{
Inline checkbox
}
