% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/custom_question_types.R
\name{question_ui_initialize.learnr_radio_inline}
\alias{question_ui_initialize.learnr_radio_inline}
\title{Inline radio buttons}
\usage{
question_ui_initialize.learnr_radio_inline(question, value, ...)
}
\description{
Inline radio buttons
}
