% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/include_etude.R
\name{include_etude}
\alias{include_etude}
\title{Inserting exercises into a document.}
\usage{
include_etude(fname, title)
}
\arguments{
\item{fname}{The path name for the etude \code{.Rmd} file.}

\item{title}{Character string containing the immediate header
for the etude, e.g., \code{"Problem 17.3:"} or \code{"### Exer A\\n\\n"}}
}
\description{
Inserts an etude into a mother document
}
