% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/show_answers.R
\name{show_answers}
\alias{show_answers}
\title{Show solutions in an etude document}
\usage{
show_answers(
  flag = TRUE,
  format = ifelse(knitr::is_latex_output(), "latex", "html")
)
}
\arguments{
\item{flag}{logical. If \code{TRUE} (default) show the chunks
marked with the \code{etude} engine as specially formatted solutions}

\item{format}{Whether the document is intended for translation
to latex or to html. Determined automatically by default.}
}
\description{
Calling \code{show_answers(TRUE)} causes the etude document
to display answers. Answers can be marked in either of two
ways:
}
\details{
\enumerate{
\item a chunk with engine \code{etude}
\item an inline chunk calling \code{ans(your_text)}
}

When \code{show_answers(TRUE)} has been called, the multiple choice questions
created by \code{true_and_false()} or \code{choose_one()} will, **if they are
being rendered as static HTML, will show the answers along with each multiple-choice
item. Note that if the document is compiled to \code{learnr}, those static
answers won't be shown. Instead, it's the user's dynamic interaction
with the \code{learnr} document that sets whether the user sees answers.
}
